renamed 'stages' pipeline to 'RAG_stages' pipeline
