+++ repository_directory_structure.txt
@@ -37,7 +37,7 @@ Repository Root: /Users/gianmariatroiani/Documents/knologi/graph_database
 │   │   ├── cosmos_db_client.py (6.6KB)
 │   │   └── document_linker.py (8.9KB)
 │   │
-│   └── stages/ [RAG Pipeline Components]
+│   └── RAG_stages/ [RAG Pipeline Components]
 │       ├── __init__.py (378B)
 │       ├── acceleration_utils.py (3.8KB)
 │       ├── chunk_text.py (19KB)
@@ -106,7 +106,7 @@ Directory Summary:
    • graph_pipeline.py - Main graph database pipeline
    • check_pipeline_setup.py - Diagnostic script for troubleshooting
    • graph_stages/ - Graph database processing modules
-   • stages/ - RAG (Retrieval-Augmented Generation) modules
+   • RAG_stages/ - RAG (Retrieval-Augmented Generation) modules
 
 📄 Document Organization:
    • global/City Comissions 2024/Agendas/ - Agenda PDF files
