+++ scripts/RAG_stages/acceleration_utils.py
